refactor(server): split agent-ops.ts and http-server.ts into sub-modules (#614)
- agent-ops.ts (787→171줄): agent-ops-core.ts(426줄)와 agent-ops-demo.ts(213줄)로 분해
- http-server.ts (754→479줄): http-server-lifecycle.ts(130줄)와 http-server-websocket.ts(134줄)로 분해
- 공개 export surface (cleanup, startServer, __test, reasonGuide, runAgentOpsCommand) 유지
- composition 패턴: 오케스트레이터가 sub-module을 import하여 기능 조합
